Updates to the MacOS install modules

+++ bin/modules/Mcol/Install/MacOS.pm
@@ -17,20 +17,40 @@ my $bin = abs_path(dirname(__FILE__) . '/../../../');
 my $applicationRoot = abs_path(dirname($bin));
 my @systemDependencies = qw(
     intltool autoconf automake expect gcc pcre2 curl libiconv pkg-config
-    openssl@3.0 mysql-client oniguruma libxml2 icu4c imagemagick mysql
-    libsodium libzip glib webp go cpanminus redis
+    openssl@3 mysql-client oniguruma libxml2 libxslt icu4c imagemagick mysql
+    libsodium libzip glib webp go cpanminus redis python@3.12
 );
 
 # ====================================
 # Subroutines
 # ====================================
 
+sub _brew_prefix {
+    my ($formula) = @_;
+    my $p = `brew --prefix $formula 2>/dev/null`; chomp $p;
+    return $p if $p;
+    $p = `brew --prefix 2>/dev/null`; chomp $p;
+    return $p || '/usr/local';
+}
+
+sub _export_brew_env {
+    my @keg = qw(openssl@3 icu4c libxml2 libzip oniguruma libxslt libiconv);
+    my @pc  = map { _brew_prefix($_).'/lib/pkgconfig' } @keg;
+    my @inc = map { _brew_prefix($_).'/include'       } @keg;
+    my @lib = map { _brew_prefix($_).'/lib'           } @keg;
+    $ENV{PKG_CONFIG_PATH} = join(':', (grep {-d} @pc), split(':', $ENV{PKG_CONFIG_PATH} // ''));
+    $ENV{CPPFLAGS} = join(' ', (map { "-I$_" } grep {-d} @inc), split(' ', $ENV{CPPFLAGS} // ''));
+    $ENV{LDFLAGS}  = join(' ', (map { "-L$_" } grep {-d} @lib), split(' ', $ENV{LDFLAGS}  // ''));
+}
+
 # Installs OS level system dependencies.
 sub install_system_dependencies {
     print "Brew is required for updating and installing system dependencies.\n";
 
     # Update/Upgrade Homebrew
-    system('brew', 'upgrade');
+    system('brew','update');
+    command_result($?, $!, "Updated Homebrew...", ['brew','update']);
+    system('brew','upgrade');
     command_result($?, $!, "Updated system dependencies...");
 
     # Filter system dependencies to only install what's missing
@@ -52,6 +72,7 @@ sub install_system_dependencies {
         print "All system dependencies already installed.\n";
     }
 
+    _export_brew_env();
     install_pip();
     install_supervisor();
 }
@@ -61,7 +82,8 @@ sub install_php {
     my ($dir) = @_;
     my $threads = how_many_threads_should_i_use();
 
-    $ENV{'PKG_CONFIG_PATH'} = '/usr/local/opt/icu4c/lib/pkgconfig';
+    _export_brew_env();
+    my $iconv_prefix = _brew_prefix('libiconv');
 
     my @configurePhp = (
         './configure',
@@ -75,17 +97,19 @@ sub install_php {
         '--enable-soap', '--enable-sockets', '--without-sqlite3',
         '--without-pdo-sqlite', '--with-libxml', '--with-xsl', '--with-zlib',
         '--with-curl', '--with-webp', '--with-openssl', '--with-zip', '--with-bz2',
-        '--with-sodium', '--with-mysqli', '--with-pdo-mysql', '--with-mysql-sock',
-        '--with-iconv=/usr/local/opt/libiconv'
+        '--with-sodium', '--with-mysqli', '--with-pdo-mysql',
+        '--with-iconv=' . $iconv_prefix
     );
 
     my $originalDir = getcwd();
 
     # Unpack PHP Archive
-    system('bash', '-c', "tar -xzf $dir/opt/php-*.tar.gz -C $dir/opt/");
-    command_result($?, $!, 'Unpacked PHP Archive...', 'tar -xf ' . $dir . '/opt/php-*.tar.gz -C ' . $dir . '/opt/');
+    system('bash', '-lc', "tar -xzf $dir/opt/php-*.tar.gz -C $dir/opt/");
+    command_result($?, $!, 'Unpacked PHP Archive...', 'tar -xzf ' . $dir . '/opt/php-*.tar.gz -C ' . $dir . '/opt/');
 
-    chdir glob("$dir/opt/php-*/");
+    my @php_dirs = glob("$dir/opt/php-*/");
+    die "PHP source not found under $dir/opt/php-*/\n" unless @php_dirs;
+    chdir $php_dirs[0];
 
     # Configure PHP
     system(@configurePhp);
@@ -98,7 +122,7 @@ sub install_php {
     print "Running make using $threads threads in concurrency.\n\n";
 
     system('make', "-j$threads");
-    command_result($?, $!, 'Compile PHP...', "make -j$threads";
+    command_result($?, $!, 'Compile PHP...', "make -j$threads");
 
     system('make install');
     command_result($?, $!, 'Installed PHP...', 'make install');
@@ -115,19 +139,18 @@ sub install_pip {
         return;
     }
 
-    print "Installing pip using get-pip.py...\n";
-
-    my $pipInstallScript = 'get-pip.py';
-    system("curl -sS https://bootstrap.pypa.io/$pipInstallScript -o $pipInstallScript");
-    command_result($?, $!, "Downloaded pip installer...");
-
-    system("chmod +x $pipInstallScript");
-    command_result($?, $!, "Made pip installer executable...");
-
-    system("python3 $pipInstallScript");
-    command_result($?, $!, "Installed pip...");
-
-    unlink($pipInstallScript);
+    # Try ensurepip first (works with Homebrew python too), fall back to get-pip.py
+    print "Installing pip via ensurepip…\n";
+    system('python3','-m','ensurepip','--upgrade');
+    if ($? != 0) {
+        print "ensurepip failed; falling back to get-pip.py…\n";
+        my $pipInstallScript = 'get-pip.py';
+        system("curl -fsSL https://bootstrap.pypa.io/get-pip.py -o $pipInstallScript");
+        command_result($?, $!, "Downloaded pip installer...");
+        system("python3 $pipInstallScript");
+        command_result($?, $!, "Installed pip...");
+        unlink($pipInstallScript);
+    }
 }
 
 # Installs Supervisor using pip, if not already installed.
@@ -140,52 +163,34 @@ sub install_supervisor {
     }
 
     print "Installing Supervisor via pip...\n";
-    system('python3', '-m', 'pip', 'install', 'supervisor');
-    command_result($?, $!, "Installed Supervisor...", 'python3', '-m', 'pip', 'install', 'supervisor');
+    system('python3', '-m', 'pip', 'install', '--user', 'supervisor');
+    command_result($?, $!, "Installed Supervisor...", 'python3 -m pip install --user supervisor');
+
+    # Ensure binaries are reachable via your app bin
+    chomp(my $user_base = `python3 -m site --user-base`);
+    my $user_bin = "$user_base/bin";
+    for my $exe (qw(supervisord supervisorctl)) {
+        my $src = "$user_bin/$exe";
+        my $dst = "$dir/bin/$exe";
+        if (-x $src && !-e $dst) { symlink $src, $dst or warn "symlink $src -> $dst failed: $!"; }
+    }
 }
 
 # installs Bazelisk.
 sub install_bazelisk {
     my ($dir) = @_;
-    my ($sysname, $nodename, $release, $version, $machine) = uname();
-    my $arch = ($machine eq 'arm64') ? 'arm64' : 'amd64';
-    my $originalDir = getcwd();
-    my $bazeliskDir = "$dir/opt/bazelisk/";
-
-    # If elixir directory exists, delete it.
-    if (-d $bazeliskDir) {
-        system(('bash', '-c', "rm -rf $bazeliskDir"));
-        command_result($?, $!, "Removing existing Bazelisk directory...", "rm -rf $bazeliskDir");
-    }
-
-    # Unpack
-    system(('bash', '-c', "tar -xzf $dir/opt/bazelisk-*.tar.gz -C $dir/opt/"));
-    command_result($?, $!, 'Unpack Bazelisk...', "tar -xzf $dir/opt/bazelisk-*.tar.gz -C $dir/opt/");
-
-    # Rename
-    system(('bash', '-c', "mv $dir/opt/bazelisk-*/ $bazeliskDir"));
-    command_result($?, $!, 'Renaming Bazelisk Dir...', "mv -xzf $dir/opt/bazelisk-*/ $bazeliskDir");
-
-    chdir glob($bazeliskDir);
-
-    # Install Bazelisk
     print "\n=================================================================\n";
-    print " Installing Bazelisk....\n";
+    print " Installing Bazelisk…\n";
     print "=================================================================\n\n";
-
-    # Install
-    system('bash', '-c', 'go install github.com/bazelbuild/bazelisk@latest');
-    command_result($?, $!, 'Install Bazelisk...', 'go install github.com/bazelbuild/bazelisk@latest');
-
-    # Binary
-    my $buildCmd = "GOOS=darwin GOARCH=$arch go build -o $dir/bin/bazel";
-    system('bash', '-c', $buildCmd);
-    command_result($?, $!, 'Build Bazelisk...', $buildCmd);
-
-    system('bash', '-c', "$dir/bin/bazel version");
+    local $ENV{GOBIN} = "$dir/bin";
+    system('go','install','github.com/bazelbuild/bazelisk@latest');
+    command_result($?, $!, 'Install Bazelisk...', ['go','install','github.com/bazelbuild/bazelisk@latest']);
+    # Provide bazel alias
+    if (-x "$dir/bin/bazelisk" && !-e "$dir/bin/bazel") {
+        symlink "$dir/bin/bazelisk", "$dir/bin/bazel" or warn "symlink bazelisk->bazel failed: $!";
+    }
+    system("$dir/bin/bazel","version");
     command_result($?, $!, 'Run Bazelisk...', "$dir/bin/bazel version");
-
-    chdir $originalDir;
 }
 
 1;
